+++ packages/terra-core-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Migrate docs from terra-toggle-section-header.
+
 ## 1.12.2 - (January 4, 2022)
 
 * Added
