+++ packages/terra-core-docs/src/terra-dev-site/doc/toggle-section-header/ToggleSectionHeader.1.doc.mdx
@@ -1,8 +1,8 @@
 import { Badge } from 'terra-toggle-section-header/package.json?dev-site-package';
 
-import DefaultToggleSectionHeader from '../example/DefaultToggleSectionHeader?dev-site-example';
-import AnimatedToggleSectionHeader from '../example/AnimatedToggleSectionHeader?dev-site-example';
-import TransparentToggleSectionHeader from '../example/TransparentToggleSectionHeader?dev-site-example';
+import DefaultToggleSectionHeader from './example/DefaultToggleSectionHeader?dev-site-example';
+import AnimatedToggleSectionHeader from './example/AnimatedToggleSectionHeader?dev-site-example';
+import TransparentToggleSectionHeader from './example/TransparentToggleSectionHeader?dev-site-example';
 
 import ToggleSectionHeaderPropsTable from 'terra-toggle-section-header/src/ToggleSectionHeader?dev-site-props-table';
