+++ packages/terra-core-docs/src/terra-dev-site/doc/toggle-section-header/example/ToggleSectionHeaderExampleTemplate.jsx
