+++ packages/terra-core-docs/src/terra-dev-site/doc/toggle-section-header/example/TransparentToggleSectionHeader.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import ToggleSectionHeaderExampleTemplate from 'terra-toggle-section-header/lib/terra-dev-site/doc/example/ToggleSectionHeaderExampleTemplate';
+import ToggleSectionHeaderExampleTemplate from './ToggleSectionHeaderExampleTemplate';
 
 const componentProps = {
-  title: 'Animated Example',
-  isAnimated: true,
+  title: 'Transparent Example',
+  isTransparent: true,
 };
 
-const AnimatedToggleSectionHeader = () => (
+const DefaultToggleSectionHeader = () => (
   <ToggleSectionHeaderExampleTemplate exampleProps={componentProps}>
     <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
     <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
@@ -14,4 +14,4 @@ const AnimatedToggleSectionHeader = () => (
   </ToggleSectionHeaderExampleTemplate>
 );
 
-export default AnimatedToggleSectionHeader;
+export default DefaultToggleSectionHeader;
